Dzień 2.2 (Handlarz fxml)(poprawka)(poprawka poprawki)

+++ src/main/java/com/example/gra/Geralt.java
@@ -9,7 +9,7 @@ public class Geralt {
     public static double power = 10;
     public static int money;
     public static double CurrentHP = 100;
-    public static double MaxHP = 100;
+    public static int MaxHP = 100;
 
     public static double onNormalAttack(ActionEvent actionEvent) throws IOException {
         Random random = new Random();
